requirements.txt
omegaconf>=2.0.6
hydra-core>=1.0.0
nltk>=3.5
numpy>=1.19.2
torch>=1.7.1
gensim>=4.0.1